html/semantics/forms/the-input-element/url.html

<!DOCTYPE html>
<html>
<head>
	<title>Input url</title>
	<link rel="author" title="Hyeonseok Shin" href="mailto:hyeonseok@gmail.com">
	<link rel="help" href="http://www.w3.org/html/wg/drafts/html/master/forms.html#url-state-%28type=url%29">
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
</head>
<body>
	<h1>Input url</h1>
	<input type="url" id="novalue" />
	<input type="url" id="value_with_LF" value="a&#x000A;a" />
	<input type="url" id="value_with_CR" value="a&#x000D;a" />
	<input type="url" id="value_with_CRLF" value="a&#x000A;&#x000D;a" />
	<input type="url" id="value_with_leading_trailing_white_space" value=" aa " />
	<input type="url" id="value_with_leading_trailing_inner_white_space" value=" a a " />
	<div id="log">
	</div>

	<script type="text/javascript">
		var element = document.getElementById('novalue');
		test(function(){
			assert_equals(element.type, 'url');
		}, 'url type supported on input element');
		test(function(){
			element.value = 'a\u000Aa';
			assert_equals(element.value, 'aa');
		}, 'User agents must not allow users to insert "LF" (U+000A)');
		test(function(){
			element.value = 'a\u000Da';
			assert_equals(element.value, 'aa');
		}, 'User agents must not allow users to insert "CR" (U+000D)');

		element = document.getElementById('value_with_LF');
		test(function(){
			assert_equals(element.value, 'aa');
		}, 'The value attribute, if specified, must have a value that contains no "LF" (U+000A)');

		element = document.getElementById('value_with_CR');
		test(function(){
			assert_equals(element.value, 'aa');
		}, 'The value attribute, if specified, must have a value that contains no "CR" (U+000D)');

		test(function(){
			element = document.getElementById('novalue');			
			element.value = 'a\u000D\u000Aa';
			assert_equals(element.value, 'aa');

			element = document.getElementById('value_with_CRLF');
			assert_equals(element.value, 'aa');
		}, 'The value sanitization algorithm is as follows: Strip line breaks from the value');

		test(function(){
			element = document.getElementById('novalue');
			element.value = ' aa ';
			assert_equals(element.value, 'aa');

			element.value = ' a a ';
			assert_equals(element.value, 'a a');

			element = document.getElementById('value_with_leading_trailing_white_space');
			assert_equals(element.value, 'aa');

			element = document.getElementById('value_with_leading_trailing_inner_white_space');
			assert_equals(element.value, 'a a');
		}, 'The value sanitization algorithm is as follows: Strip leading and trailing whitespace from the value.');
	</script>

</body>
</html>
