LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/url-expected.txt

Input url


PASS url type supported on input element 
PASS The value must not be set with "LF" (U+000A) or "CR" (U+000D) 
PASS The value sanitization algorithm is as follows: Strip line breaks from the value 
FAIL The value sanitization algorithm is as follows: Strip leading and trailing whitespace from the value. assert_equals: expected "aa" but got " aa "
